Finished login and logout process
