+++ app/routes/account.js
@@ -4,11 +4,4 @@ import { inject as service } from '@ember/service';
 export default class LoginRoute extends Route {
   @service session;
   @service router;
-
-  beforeModel() {
-    this.session.prohibitAuthentication('home');
-    if (this.session.data.authenticated.token !== undefined) {
-      this.router.transitionTo('account');
-    }
-  }
 }
