+++ app/templates/account.hbs
@@ -1,0 +1,9 @@
+<div class="account">
+    <h1>Please login or logout of your account here!</h1>
+    <div class="account-buttons">
+        <button {{on "click" this.logout}}>Logout</button>
+    <button><LinkTo @route="login" class="menu-login">
+        Login
+    </LinkTo></button>
+    </div>
+</div>
