+++ app/templates/login.hbs
@@ -2,6 +2,7 @@
     <p class="error"><strong>{{this.error}}</strong></p>
 {{/if}}
 <div class="login">
+    <br>
     <p>No Account: </p><LinkTo @route="register" class="button" >Register Here</LinkTo>
     <br>
     <h1>Log into your account</h1>
